Update to create-react-app v2.1.0
