--- junit_html_report/packages/office-ui-fabric-react-tslint/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "office-ui-fabric-react-tslint",
-  "version": "5.0.0",
-  "description": "Shared tslint rules",
-  "main": "tslint.json",
-  "scripts": {
-    "clean": "",
-    "build": "node -v"
-  },
-  "devDependencies": {
-    "tslint-react": "^3.2.0"
-  }
-}
