+++ junit_html_report/src/index.css
@@ -3,3 +3,7 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+.GroupedList-name {
+  padding-bottom: 40px;
+}
